Do not parse numbers that evaluate to Infinity
Some numbers, like `87e3202`, are recognised as numbers by the
`isNumber` utility, which results on minimist parsing them by default
and thus passing a meaningless `Infinity` value to the user.

Since parsing scientific notation numbers are supported on minimist as a
feature, the approach of this patch is to only return `true` from
`isNumber` if the string indeed represents a number, but it doesn't
resolve to `Infinity` when being parsed as such, therefore not breaking
any existing functionality on the module.

The use case where this bug showed up was passing (shorter) UUIDs as
command line arguments.